requirements.txt
# app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
aiohttp==3.9.5
azure-cosmos==4.9.0
dotenv 

# --- LangChain, OpenAI y AI ---
openai==1.69.0
langchain==0.3.22
langchain-core==0.3.49
langchain-community==0.3.20
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langgraph==0.3.21
langgraph-checkpoint==2.0.23
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.60
langsmith==0.3.19
tiktoken==0.9.0
propcache==0.3.1
jiter==0.9.0

# --- Azure SDKs ---
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.21.0
azure-keyvault-secrets==4.9.0
azure-search-documents==11.5.2
azure-cosmos==4.9.0
msal==1.32.0
msal-extensions==1.3.1
azure-storage-blob==12.25.1